refactor: add the same types from frontend to backend
Eventually I'd like to use something like verdaccio or leverage
some monorepo nicities, but for now I'm going to take the
plebian route and just copy/paste
