--- apps/be/src/types/DailyForecast.ts
@@ -1,9 +1,0 @@
-type DailyForecast = {
-  // TODO: Can this come from the frontend?
-};
-
-type DateString = string & { __brand: 'DateString' };
-
-type MultiDayForecast = Record<DateString, DailyForecast>;
-
-export type { DailyForecast, MultiDayForecast };
