+++ apps/fe/src/components/weather-provider.tsx
@@ -2,7 +2,7 @@ import { createContext, useContext, useState, type ReactNode } from 'react';
 import { formatDate, getDatesForNextTwoWeeks } from '@/lib/date-utils';
 import { mockForecastData, mockHistoricalData } from '@/lib/mock-data';
 import type {
-  DailyForecast,
+  DailyForecastList,
   HistoricalData,
   WeatherContextType,
 } from '@/lib/types';
@@ -10,7 +10,7 @@ import type {
 const WeatherContext = createContext<WeatherContextType | undefined>(undefined);
 
 export function WeatherProvider({ children }: { children: ReactNode }) {
-  const [forecastData] = useState<DailyForecast[]>(mockForecastData);
+  const [forecastData] = useState<DailyForecastList>(mockForecastData);
   const [historicalData] =
     useState<Record<string, HistoricalData>>(mockHistoricalData);
   const [selectedDate, setSelectedDate] = useState<string | null>(
